unittests asc-core/src/utils 100%
