+++ packages/asc-core/src/components/ButtonStyle/ButtonStyle.ts
@@ -1,6 +1,6 @@
 import { readableColor } from 'polished'
 import styled from '../../styled-components'
-import { fillSVG, getColorFromTheme } from '../../utils'
+import { fillSvgFromTheme, getColorFromTheme } from '../../utils'
 import ButtonBaseStyle, { ButtonBaseStyleProps } from '../ButtonBaseStyle'
 
 export type Props = ButtonBaseStyleProps
@@ -25,7 +25,7 @@ const ButtonStyle = styled(ButtonBaseStyle)<Props>`
   & svg {
     width: 30px;
     height: 30px;
-    ${({ color, theme }) => fillSVG(theme, color)};
+    ${({ color, theme }) => fillSvgFromTheme(theme, color)};
   }
 `
