+++ packages/asc-core/src/utils/index.ts
@@ -1,8 +1,8 @@
 export { default as fromTheme } from './fromTheme'
-export { default as fillSVG } from './fillSVG'
 export {
   getColorFromTheme,
   getTypographyFromTheme,
   getFocusStyle,
   getBreakpointFromTheme,
+  fillSvgFromTheme,
 } from './themeUtils'
