+++ packages/asc-core/src/utils/themeUtils.ts
@@ -1,21 +1,22 @@
+import { readableColor } from 'polished'
 import { css } from '../styled-components'
 import { Theme } from '../theme'
 import { fromTheme } from '.'
 
-export function getColorFromTheme(
+export const getColorFromTheme = (
   theme: Theme.ThemeInterface,
   colorType?: Theme.TypeLevel,
   variant: string = 'main',
-) {
+) => {
   return colorType
     ? fromTheme(`colors.${[colorType]}.${[variant]}`)({ theme })
     : fromTheme('colors.tint.level1')({ theme })
 }
 
-export function getTypographyFromTheme(
+export const getTypographyFromTheme = (
   theme: Theme.ThemeInterface,
   attributeType: string,
-) {
+) => {
   return fromTheme(`typography.${[attributeType]}`)({ theme })
 }
 
@@ -26,15 +27,36 @@ export const getFocusStyle = (theme: Theme.ThemeInterface) => css`
     outline-width: medium;
   }
 `
-export function getBreakpointFromTheme(
+export const getBreakpointFromTheme = (
   theme: Theme.ThemeInterface,
   type: Theme.TypeBreakpoint,
   variant: string,
-) {
+) => {
   const breakpoint: Theme.GetBreakpointFunc = fromTheme(
     `breakpoints.${[variant]}`,
   )({
     theme,
   })
   return breakpoint && breakpoint(type)
+}
+
+export const fillSvgFromTheme = (
+  theme: Theme.ThemeInterface,
+  colorType?: Theme.TypeLevel,
+  variant: string = 'main',
+) => {
+  if (colorType) {
+    const color = getColorFromTheme(theme, colorType, variant)
+    if (typeof color === 'string') {
+      return `
+        rect,
+        polygon,
+        path {
+          fill: ${readableColor(color)}
+        }
+      `
+    }
+  }
+
+  return ''
 }
